projects: adf4030: use the example infrastructure
Migrate adf4030 to the shared examples infrastructure (examples.mk): add
EXAMPLE := basic + include examples.mk to the Makefile, move the iio
sources into src/examples/iio/example.mk (IIOD=y), drop examples_src.mk
and platform_includes.h, and shorten the iio example directory
(iio_example -> iio). main.c now defers to example_main(); the basic
example owns its UART stdio init and the iio example keeps iio_app.h.
Normalize builds.json keys/flags to EXAMPLE=iio and fix the README's
stale iio_example references.

Verified by dependency resolution (no local mbed toolchain): both
examples resolve their full SRCS/INCS and every referenced file exists.
