+++ projects/adf4030/Makefile
@@ -1,7 +1,11 @@
 # Uncomment to use the desired platform
 PLATFORM = mbed
 
+EXAMPLE ?= basic
+
 include ../../tools/scripts/generic_variables.mk
+
+include ../../tools/scripts/examples.mk
 
 include src.mk
