+++ projects/adf4030/src/examples/iio/example.mk
@@ -1,0 +1,4 @@
+IIOD = y
+
+INCS += $(DRIVERS)/frequency/adf4030/iio_adf4030.h
+SRCS += $(DRIVERS)/frequency/adf4030/iio_adf4030.c
