+++ projects/adf4030/src/platform/mbed/main.c
@@ -34,17 +34,11 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 #include "no_os_error.h"
 
-#ifdef BASIC_EXAMPLE
-#include "basic_example.h"
-#endif
-
-#ifdef IIO_EXAMPLE
-#include "iio_example.h"
-#endif
+extern int example_main();
 
 /**
  * @brief Main function execution for mbed platform.
@@ -52,24 +46,5 @@
  */
 int main()
 {
-	int ret = -EINVAL;
-
-#ifdef BASIC_EXAMPLE
-	struct no_os_uart_desc *uart_desc;
-
-	ret = no_os_uart_init(&uart_desc, &adf4030_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart_desc);
-	ret = basic_example_main();
-	if (ret)
-		no_os_uart_remove(uart_desc);
-#endif
-
-#ifdef IIO_EXAMPLE
-	ret = iio_example_main();
-#endif
-
-	return ret;
+	return example_main();
 }
